+++ tidy3d/__init__.py
@@ -539,8 +539,8 @@ __all__ = [
     "ChargeToleranceSpec",
     "DCSpec",
     "CaugheyThomasMobility",
-    "SlotboomNarrowingModel",
-    "SRHRecombination",
+    "SlotboomNarrowingBandgap",
+    "ShockleyReedHallRecombination",
     "AugerRecombination",
     "RadiativeRecombination",
     "VoltageData",
